refactor(LFXV2-2444): colocate CLI logic and drop Helm job template
Move rename-project-slug migration into cmd/project-cli/commands/sync,
rename NATS connect helpers to client.go/NatsConfig/Connect, and align
in-cluster runs with committee-cli via kubectl Job manifest docs.
