+++ cmd/project-cli/README.md
@@ -6,6 +6,8 @@ A CLI tool for running operational tasks against the project service. Commands a
 
 Operational scripts such as slug renames need a repeatable, auditable, cluster-native execution path with shared infrastructure wiring, dry-run support, structured logs, and deterministic exit codes. The CLI reuses the project-service NATS and OpenSearch client setup rather than duplicating connection logic in standalone scripts.
 
+Like `committee-cli`, **the Helm chart deploys only the API**. Operational jobs are not rendered by the chart — create a one-off Kubernetes Job when you need to run a command in-cluster.
+
 ## Usage
 
 ```text
@@ -94,48 +96,91 @@ In CI, the image is built and published automatically by the existing `ko-build-
 
 ## Running as a Kubernetes Job
 
-Enable the Helm job template and set slug parameters in `values.yaml`:
+Create a one-off Job from the published `project-cli` image. Replace `<tag>` with the image tag you want to run (for example the chart `appVersion` or a CI build tag).
 
-```yaml
-jobs:
-  renameProjectSlug:
-    enabled: true
-    oldSlug: old-slug
-    newSlug: new-slug
-    dryRun: "true"
-    target: both
-```
+### Quick start (`kubectl create job`)
 
-Then upgrade the chart or render the job manifest:
+Use this when the command only needs flags and the image defaults are acceptable for your cluster:
 
 ```sh
-helm template lfx-v2-project-service ./charts/lfx-v2-project-service \
-  --set jobs.renameProjectSlug.enabled=true \
-  --set jobs.renameProjectSlug.oldSlug=old-slug \
-  --set jobs.renameProjectSlug.newSlug=new-slug
-```
-
-Or create a one-off job from the published image:
-
-```sh
-kubectl create job lfx-project-cli-rename-slug \
+kubectl create job lfx-project-cli-rename-slug-dry-run \
   --image=ghcr.io/linuxfoundation/lfx-v2-project-service/project-cli:<tag> \
   --namespace=lfx \
-  --env="OLD_SLUG=old-slug" \
-  --env="NEW_SLUG=new-slug" \
-  --env="DRY_RUN=true" \
-  -- sync rename-project-slug
+  -- sync rename-project-slug old-slug new-slug
 ```
 
-Monitor progress:
+`kubectl create job` does not set environment variables. For in-cluster runs you typically need `NATS_URL`, `OPENSEARCH_URL`, and slug settings — use the manifest below instead.
+
+### Recommended: apply a Job manifest
+
+Save as `rename-project-slug-job.yaml`, edit slugs and URLs for your environment, then apply:
+
+```yaml
+apiVersion: batch/v1
+kind: Job
+metadata:
+  generateName: lfx-project-cli-rename-slug-
+  namespace: lfx
+spec:
+  backoffLimit: 0
+  ttlSecondsAfterFinished: 604800
+  template:
+    spec:
+      restartPolicy: Never
+      containers:
+        - name: project-cli
+          image: ghcr.io/linuxfoundation/lfx-v2-project-service/project-cli:<tag>
+          args:
+            - sync
+            - rename-project-slug
+          env:
+            - name: NATS_URL
+              value: nats://lfx-platform-nats.lfx.svc.cluster.local:4222
+            - name: OPENSEARCH_URL
+              value: http://opensearch-cluster-master.lfx.svc.cluster.local:9200
+            - name: LOG_LEVEL
+              value: info
+            - name: JOB_RUN_ID
+              valueFrom:
+                fieldRef:
+                  fieldPath: metadata.name
+            - name: OLD_SLUG
+              value: old-slug
+            - name: NEW_SLUG
+              value: new-slug
+            - name: DRY_RUN
+              value: "true"
+            - name: TARGET
+              value: both
+            - name: CONCURRENCY
+              value: "50"
+          resources:
+            limits:
+              cpu: 500m
+              memory: 512Mi
+            requests:
+              cpu: 100m
+              memory: 128Mi
+```
 
 ```sh
-kubectl logs -f job/lfx-project-cli-rename-slug -n lfx
+kubectl apply -f rename-project-slug-job.yaml
 ```
+
+Run a live migration by setting `DRY_RUN` to `"false"` (and reviewing dry-run logs first).
+
+### Monitor and re-run
+
+```sh
+kubectl get jobs -n lfx | grep project-cli-rename-slug
+kubectl logs -f job/<job-name> -n lfx
+```
+
+The Job is kept after completion so its logs and exit status remain accessible as a run record. Re-trigger by creating a new Job (new `generateName` prefix or a different job name).
 
 ## Adding new commands
 
 1. Create `cmd/project-cli/commands/<group>/` and implement the `Command` and `Subcommand` interfaces from `commands/command.go`.
 2. Register the new command in `buildRegistry()` in `cmd/project-cli/main.go`.
 
-No changes to shared infrastructure are required unless the new command needs a client that does not yet exist.
+No changes to the Helm chart are required. Document the new command here with a `kubectl create job` or Job manifest example when it is intended for in-cluster use.
